Use a float when calculating throttling

+++ src/common/bu/ResourceManager.cc
@@ -459,7 +459,7 @@ float evb::bu::ResourceManager::getAvailableResources()
       resourceLimitiationAlreadyNotified_ = true;
     }
 
-    resourcesFromFUs *= 1 - ( (lsLatency - configuration_->lumiSectionLatencyLow) /
+    resourcesFromFUs *= 1 - ( static_cast<float>(lsLatency - configuration_->lumiSectionLatencyLow) /
                               (configuration_->lumiSectionLatencyHigh - configuration_->lumiSectionLatencyLow) );
   }
